Hides excess authors
The first 10 authors are shown and `et al` added if there are more than 10.

For the Item view only, a total count of authors is displayed and the ability to hide/show
them is included.

Fixes #7

Some of this logic may be better done in the controller, but as it's been extracted out to
a partial view (i.e. method) it seems okay enough for now.