Updates all sidenavs to use list-group
The size of the sidenav div had to grow for some of these to work
properly without the badge looking weird (dropping to it’s own line and
overlapping). If we are concerned with that, I can spend some more time
on that now or we can just address it later if/when it becomes a real
estate problem. My current thought is to leave it like this and move to
a truly fluid layout eventually that’ll minimize this “problem”. I’ll
open a ticket to do that and if we decide not to we can close it.