METS generator
There are a few METS field id values that seem arbitrary so they are
left out (they are incremental div id numbers that are not reference
elsewhere for example).

File names are their own method, so the Packager should be able to use
the same method to generate a name matching the toMets method. (And
thus if we need to make a change, we should be able to just change what
that method does without too much concern.

`toMets` should be fairly forgiving and just generate a file with
whatever metadata exists for an Item that we’ve mapped to METS. As
such, the notion of a ContentProfile is not actually being checked for
this work. I’m fine with adding it back in if we decide an Analyst
really will want the flexibility to Catalog metadata without including
it in the mets.xml file.

The packager will be responsible for calling this method and writing
the actual file into the zip.

The route `/item/package/:id` is a convenience for testing this at this
time. Eventually that route will be used to delivery one-off Sword
packages to a configured repository. For now, it just dumps the mets
file.

I will update the Content Model in our private repo that extracts the
additional metadata fields this work uses (although it’s all optional
and will be excluded from output if it does not exist rather than
failing).

closes #86