Filter Search Results by Topic subscription status
This allows for filtering Search Results for Topics by Subscription status.

The default is Any status, but you can choose Only Subscribed or Only Unsubscribed options.

An annoyace remains in which the total number of results reported is inaccurate. This is because we rely on elasticsearch to report on the total resultset, but are then filtering on just a single page of results.

It is theoretically possible to refactor this to have elasticsearch do the heavy lifting by passing in TopicIDs to Include or Exclude from the search which would resolve that remaining issue.

Closes #77