Default DB URL and Fake Email Key
As a first round of work on #22, this provides default db url and a
fake email api key.

If they are defined in the ENV, the ENV values win, so Heroku and
Travis should remain fine with this config but this has the benefit of
not requiring setting those values if they are the development versions
like we already do for elasticsearch. If a developer or deployed server
desires non-default values, they can easily set via ENV.