Query elasticsearch with json
This attempts to use json as the query instead of the rest api.

The rest api appears quite limited, and will likely not be sufficient to
exclude all topics a user is subscirbed to if we decide to do that via
the actual search rather than hack it post search.

This is not complete and fairly broken.