Hold Pages Spec and Topic Pick Pages Spec
part of #128

See #170 for why some tests are pending. In short, the controller
methods call an actor which is not available in our current test
environment. It’s solve-able, but non-trivial. The controllers are
indeed protected by the appropriate roles/subscriber affiliation but
we’re not really confirming that a hold is resolved properly (or not)
as it always seems to not be resolved due to the lack of the worker.