Refactor Authentication in Test Env
- In order to move to storing a SubscriberId in the session we need to
move the test Env authentication closer to how production works. This
is actually win/win as now we are testing more of the production
authentication path as a side effect of this work. Yay.
- app/services/OAuth2.scala#currentSubscriber is really just a stub for
future functionality in which we’ll need to figure out a way to handle
Users with multiple Subscribers.