Makes RequestHeader available in all views (and layout)
We need to access some session variables in Views as well as in the
main Layout view (Navigation in particular). The implicit RequestHeader
allows for that and is the recommended way as of Play 2.3.x instead of
accessing the session directly.

If we just needed Session data it in some views, we could pass it to
them via the Controller explicitly. However, since we also need it in
the main Layout template, we seemingly need to implicitly pass it to
each view and have those views implicitly pass it to the Layout. Or at
least that’s how I solved this.