Detects unhanded collections in OAI feed
If any Collections are detected in an OAI Harvest that are not handled,
an email is sent to sysadmins (this may want to change eventually to
instead email the appropriate Analysts but for now sysadmins works well
for us).

This also adds a feature to create Inactive Collections which is
essentially stating you have no interest in the Collection which will
prevent notification when it is detected again on a future Harvest.

closes #81