Method to kick off all harvests (protected by application key)
The intent of this route and method are for a cron-like job to kick off
the process. As such, we need to handle this in a non-logged in state.
The application defined key should suffice for our level of protection
needs.

Other changes include: all Harvests default to frequency of 1 and that
cannot be changed in the UI. Admins can still kick off a harvest with a
greater frequency for catchup or testing purposes in the manual harvest
start form.