Prevent batch harvests from requesting invalid dates
This prevents batch harvests from requesting date that is older than
yesterday with the assumption that it’s safest to not request an OAI
feed for today as more data may be added later we’d miss. I’m not
entirely sure that’s valid logic, but this is safe. The harvester
increments the updated date hence the need to make sure it is less than
yesterday (i.e. updated needs to be two days ago to actually run which
would then get incremented by 1 to be a harvest for yesterday… clear as
mud).

Admin kicked off harvests and knips are not limited by date. Only the
batch starter.

closes #240