METS author limits, date created, and TopicHub as agent
This allows setting METS_RESTRICT_MAXIMUM_AUTHORS to an integer. If
there are more authors than that number, none are supplied in the METS
to prevent triggering a bug in our local DSpace installation.

Fixes a bug where we included a title instead of the date.

Adds dc.agent as a mechanism for detecting the package came from
TopicHub. If a better field is identified, this could be changed and we
can work with Carl to update the crosswalk that relies on this value.

closes #371
closes #370
closes #375